chore: Convert index.js to index.mjs
